OCRmypdf_plugins/ocrmypdf-chandra-main/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ocrmypdf-chandra"
version = "0.1.0"
description = "Chandra OCR plugin for OCRmyPDF"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MPL-2.0" }
authors = [{ name = "img_2_pdf contributors" }]
dependencies = [
  "ocrmypdf>=17.0.0,<18.0.0",
  "chandra-ocr[hf]>=0.2.0",
  "pillow>=10.0.0",
]

[tool.setuptools]
packages = ["ocrmypdf_chandra"]
